+++ app/webFrontend/src/app/shared/services/data.service.ts
@@ -492,9 +492,15 @@ export class DownloadFileService extends DataService {
     super('download/', backendService);
   }
 
-  postDownloadReqForCourse(idl: IDownload): Promise<Response> {
+  postDownloadReqForCoursePDFIndividual(idl: IDownload): Promise<Response> {
     return this.backendService
-      .post(this.apiPath, idl)
+      .post(this.apiPath + 'pdf/individual', idl)
+      .toPromise();
+  }
+
+  postDownloadReqForCoursePDFSingle(idl: IDownload): Promise<Response> {
+    return this.backendService
+      .post(this.apiPath + 'pdf/single', idl)
       .toPromise();
   }